requirements.txt
Django==3.1.6
python-telegram-bot==13.2